requirements.txt
Django==3.1.7
djangorestframework==3.12.4
pytz==2021.1
mysqlclient==2.1.0
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0
djangorestframework-camel-case==1.2.0
django-url-filter==0.3.15  # https://github.com/miki725/django-url-filter
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
drf-friendly-errors-2==0.14.3  # https://pypi.org/project/drf-friendly-errors-2/